requirements.txt
pandas>=2.1.0
numpy>=1.26.0
mysql-connector-python>=8.2.0
python-dotenv>=1.0.0
